requirements.txt
absl-py==2.2.1
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
attrs==25.3.0
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chumpy==0.70
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
Comet==3.1.0
comet_ml==3.49.5
configobj==5.0.9
constantly==23.10.4
contourpy==1.3.1
coverage==7.8.0
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
Cython==3.0.12
dlib==19.24.6
dulwich==0.22.8
everett==3.1.0
filelock==3.14.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
humanfriendly==10.0
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
incremental==24.7.2
iniconfig==2.1.0
jax==0.5.3
jaxlib==0.5.3
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.4.2
json-tricks==3.17.3
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
lmdb==1.6.2
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
mmengine==0.10.7
mmpose==1.3.2
model-index==0.1.11
mpmath==1.3.0
munkres==1.1.4
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.0
numpy==1.26.4
onnx==1.17.0
onnxoptimizer==0.3.13
onnxruntime==1.21.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
opt_einsum==3.4.0
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
protobuf==4.25.6
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pycryptodome==3.22.0
Pygments==2.19.1
PyMatting==1.1.13
pyparsing==3.2.1
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
pytest==8.3.5
pytest-runner==6.0.1
python-box==6.1.0
python-dateutil==2.9.0.post0
pytorch-metric-learning==2.8.1
PyTurboJPEG==1.7.7
pytz==2023.4
PyYAML==6.0.2
referencing==0.36.2
rembg==2.0.64
requests==2.28.2
requests-toolbelt==1.0.0
rich==13.4.2
rpds-py==0.23.1
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2

torch==2.6.0
torchvision==0.21.0
tqdm==4.65.2
ultralytics==8.3.148
xtcocotools==1.14.3
yaml-config-override==0.5
